editing
edit some bugs in delete and some responsive in index head

+++ admin-panel/blogs/blog_delete.php
@@ -4,8 +4,9 @@ require_once('../../app/loader.php');
 if($_SERVER['REQUEST_METHOD'] == 'GET'){
     $id = securityCheck($_REQUEST['id']);
     $path = $db->where('id', $id)
-    ->getOne('ads', 'image'); 
-    unlink('../../'.$path['image']);
+    ->getOne('blogs', 'image'); 
+    if(file_exists('../../'.$path['image']))
+        unlink('../../'.$path['image']);
     $db->where('id', $id)
     ->delete('blogs');
