+++ admin-panel/slideshow/slideshow_delete.php
@@ -5,8 +5,9 @@ if($_SERVER['REQUEST_METHOD'] == 'GET'){
     $id = securityCheck($_REQUEST['id']);
     
     $path = $db->where('id', $id)
-    ->getOne('ads', 'image'); 
-    unlink('../../'.$path['image']);
+    ->getOne('slideshows', 'image'); 
+    if(file_exists('../../'.$path['image']))
+        unlink('../../'.$path['image']);
     $db->where('id', $id)
     ->delete('slideshows');
